create a lsp transport reader wrapper
